+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
-* Add a code formatter for python files
-* Add support for new classifiers and mappers
+* Added a code formatter for python files.
+* Added support for new classifiers and mappers.
 
 #### 1.0.10
 * Fixed error where changes to only non-validated files would fail validation.